requirements.txt
Django==3.2
djangorestframework==3.13.1
djangorestframework_simplejwt==5.0.0
Pillow==9.0.0